revamped contact layout, imagery
